requirements.txt
requests
beautifulsoup4
lxml
playwright
python-dotenv
loguru
icalendar
